Project-Hospital-main 2/requirements.txt
pymongo
pydgraph
cassandra-driver ==3.28.0
time_uuid